# schememitra/packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SchemeMitra Core API",
    "description": "Deterministic eligibility, repayment maths, partner routing, documents, voice and application tracking for marginalized entrepreneurs. Money is integer paise; rates are bps.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/auth/otp/request": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Otp Request",
        "operationId": "otp_request_v1_auth_otp_request_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OtpRequestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/otp/verify": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Otp Verify",
        "operationId": "otp_verify_v1_auth_otp_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpVerify"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/eligibility/evaluate": {
      "post": {
        "tags": [
          "eligibility"
        ],
        "summary": "Evaluate",
        "description": "Rules-only eligibility with a full, localisable trace for every scheme.",
        "operationId": "evaluate_v1_eligibility_evaluate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/documents": {
      "post": {
        "tags": [
          "documents"
        ],
        "summary": "Upload Document",
        "operationId": "upload_document_v1_documents_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_document_v1_documents_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/documents/digilocker/consent": {
      "post": {
        "tags": [
          "documents"
        ],
        "summary": "Digilocker Consent",
        "description": "Exchanges the DigiLocker authorization code for a short-lived access token.",
        "operationId": "digilocker_consent_v1_documents_digilocker_consent_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsentTokenIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Digilocker Consent V1 Documents Digilocker Consent Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/documents/digilocker/authorize": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "Digilocker Authorize",
        "operationId": "digilocker_authorize_v1_documents_digilocker_authorize_get",
        "parameters": [
          {
            "name": "state",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "State"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Digilocker Authorize V1 Documents Digilocker Authorize Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/documents/digilocker/pull": {
      "post": {
        "tags": [
          "documents"
        ],
        "summary": "Digilocker Pull",
        "operationId": "digilocker_pull_v1_documents_digilocker_pull_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PullIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/documents/{document_id}/image": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "Document Image",
        "operationId": "document_image_v1_documents__document_id__image_get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/demo/documents/{persona}/{doc_type}": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "Demo Document",
        "description": "Fictional demo papers for judges (DEMO_MODE only).",
        "operationId": "demo_document_v1_demo_documents__persona___doc_type__get",
        "parameters": [
          {
            "name": "persona",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-z]+$",
              "title": "Persona"
            }
          },
          {
            "name": "doc_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-z_]+$",
              "title": "Doc Type"
            }
          },
          {
            "name": "variant",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "paper",
                "digilocker"
              ],
              "type": "string",
              "default": "paper",
              "title": "Variant"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/demo/documents/{persona}": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "Demo Document List",
        "operationId": "demo_document_list_v1_demo_documents__persona__get",
        "parameters": [
          {
            "name": "persona",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-z]+$",
              "title": "Persona"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "title": "Response Demo Document List V1 Demo Documents  Persona  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_document_v1_documents_post": {
        "properties": {
          "application_id": {
            "type": "string",
            "format": "uuid",
            "title": "Application Id"
          },
          "doc_type": {
            "type": "string",
            "maxLength": 32,
            "title": "Doc Type"
          },
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "source": {
            "type": "string",
            "enum": [
              "camera",
              "upload"
            ],
            "title": "Source",
            "default": "camera"
          }
        },
        "type": "object",
        "required": [
          "application_id",
          "doc_type",
          "file"
        ],
        "title": "Body_upload_document_v1_documents_post"
      },
...
      "ConsentTokenIn": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          }
        },
        "type": "object",
        "required": [
          "code"
        ],
        "title": "ConsentTokenIn"
      },
...
      "DocumentOut": {
        "properties": {
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "fields": {
            "additionalProperties": true,
            "type": "object",
            "title": "Fields"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "issues": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Issues"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "suggest_digilocker": {
            "type": "boolean",
            "title": "Suggest Digilocker"
          },
          "verified": {
            "type": "boolean",
            "title": "Verified"
          },
          "sandbox": {
            "type": "boolean",
            "title": "Sandbox"
          },
          "engine": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Engine"
          },
          "sha256": {
            "type": "string",
            "title": "Sha256"
          },
          "purge_after": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Purge After"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "type",
          "source",
          "fields",
          "confidence",
          "issues",
          "status",
          "suggest_digilocker",
          "verified",
          "sandbox",
          "engine",
          "sha256",
          "purge_after"
        ],
        "title": "DocumentOut"
      },
...
      "EvaluateRequest": {
        "properties": {
          "applicant": {
            "$ref": "#/components/schemas/ApplicantFacts"
          },
          "scheme_codes": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheme Codes"
          }
        },
        "type": "object",
        "required": [
          "applicant"
        ],
        "title": "EvaluateRequest"
      },
      "EvaluateResponse": {
        "properties": {
          "rules_etag": {
            "type": "string",
            "title": "Rules Etag"
          },
          "eligible": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Eligible"
          },
          "ineligible": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Ineligible"
          },
          "needs_info": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Needs Info"
          },
          "near_misses": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Near Misses"
          },
          "next_best": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Best"
          },
          "ranking_method": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "model",
                  "heuristic",
                  "none"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Ranking Method"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/SchemeResult"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "rules_etag",
          "eligible",
          "ineligible",
          "needs_info",
          "near_misses",
          "next_best",
          "results"
        ],
        "title": "EvaluateResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OtpRequest": {
        "properties": {
          "phone": {
            "type": "string",
            "maxLength": 16,
            "minLength": 10,
            "title": "Phone",
            "examples": [
              "9876500001"
            ]
          }
        },
        "type": "object",
        "required": [
          "phone"
        ],
        "title": "OtpRequest"
      },
      "OtpRequestResponse": {
        "properties": {
          "sent": {
            "type": "boolean",
            "title": "Sent"
          },
          "ttl_seconds": {
            "type": "integer",
            "title": "Ttl Seconds"
          },
          "channel": {
            "type": "string",
            "title": "Channel"
          },
          "dev_otp": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dev Otp"
          }
        },
        "type": "object",
        "required": [
          "sent",
          "ttl_seconds",
          "channel"
        ],
        "title": "OtpRequestResponse"
      },
      "OtpVerify": {
        "properties": {
          "phone": {
            "type": "string",
            "maxLength": 16,
            "minLength": 10,
            "title": "Phone"
          },
          "code": {
            "type": "string",
            "maxLength": 8,
            "minLength": 4,
            "title": "Code"
          },
          "lang": {
            "type": "string",
            "maxLength": 8,
            "title": "Lang",
            "default": "en"
          }
        },
        "type": "object",
        "required": [
          "phone",
          "code"
        ],
        "title": "OtpVerify"
      },
...
      "PullIn": {
        "properties": {
          "application_id": {
            "type": "string",
            "format": "uuid",
            "title": "Application Id"
          },
          "doc_type": {
            "type": "string",
            "title": "Doc Type"
          },
          "consent_token": {
            "type": "string",
            "title": "Consent Token"
          }
        },
        "type": "object",
        "required": [
          "application_id",
          "doc_type",
          "consent_token"
        ],
        "title": "PullIn"
      },
...
      "SchemeResult": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Name"
          },
          "apex_corp": {
            "type": "string",
            "title": "Apex Corp"
          },
          "loan_type": {
            "type": "string",
            "title": "Loan Type"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "rule_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rule Version Id"
          },
          "source_url": {
            "type": "string",
            "title": "Source Url"
          },
          "verified_on": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verified On"
          },
          "needs_verification": {
            "type": "boolean",
            "title": "Needs Verification"
          },
          "status": {
            "type": "string",
            "enum": [
              "eligible",
              "ineligible",
              "needs_info"
            ],
            "title": "Status"
          },
          "trace": {
            "items": {
              "$ref": "#/components/schemas/TraceRow"
            },
            "type": "array",
            "title": "Trace"
          },
          "failed": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Failed"
          },
          "missing": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing"
          },
          "near_miss": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TraceRow"
              },
              {
                "type": "null"
              }
            ]
          },
          "documents_required": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Documents Required"
          },
          "channel_partner_types": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Channel Partner Types"
          },
          "offer": {
            "$ref": "#/components/schemas/Offer"
          },
          "rank": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RankInfo"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "code",
          "name",
          "apex_corp",
          "loan_type",
          "version",
          "rule_version_id",
          "source_url",
          "verified_on",
          "needs_verification",
          "status",
          "trace",
          "failed",
          "missing",
          "near_miss",
          "documents_required",
          "channel_partner_types",
          "offer"
        ],
        "title": "SchemeResult"
      },
...
      "SessionUser": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "partner_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Partner Id"
          },
          "csc_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Csc Id"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "preferred_lang": {
            "type": "string",
            "title": "Preferred Lang",
            "default": "en"
          }
        },
        "type": "object",
        "required": [
          "id",
          "role"
        ],
        "title": "SessionUser"
      },
...
      "TokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In"
          },
          "user": {
            "$ref": "#/components/schemas/SessionUser"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "token_type",
          "expires_in",
          "user"
        ],
        "title": "TokenResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}